Restructure into feature packages: register/ package, ARCHITECTURE.md, and a subpackaged test tree
