+++ crudauth/register/__init__.py
@@ -1,0 +1,7 @@
+"""Self-registration feature: the ``/register`` route and its request body."""
+
+from __future__ import annotations
+
+from .route import RegisterIn, build_register_route
+
+__all__ = ["RegisterIn", "build_register_route"]
