+++ tests/core/test_column_map.py
