+++ tests/core/test_logical_fields.py
