+++ tests/register/test_gating.py
