+++ tests/register/test_route.py
