+++ tests/transports/test_session_devices.py
